requirements.txt
ccxt
pandas
ta
python-telegram-bot==13.15